# api/openapi.yaml
openapi: 3.1.0
info:
  title: ansible-automation-portal
  version: '1'
  description: |
    API for the Ansible Automation Portal backend plugin in Red Hat Developer Hub / Backstage.
    Provides endpoints for syncing AAP resources into the Backstage catalog, managing execution
    environments, and proxying Git SCM operations.
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact: {}
servers:
  - url: /
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /apme/settings:
    get:
      operationId: getApmeSettings
      summary: Get APME portal settings
      description: Returns feature flags for AI and remediation publish mode configured for the portal.
      tags:
        - APME
      responses:
        '200':
          description: APME settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  enableAi:
                    type: boolean
                  publishViaGateway:
                    type: boolean
                  targetAnsibleCoreVersion:
                    type: string
                    description: Global default ansible-core minor version for quality scans
    put:
      operationId: updateApmeSettings
      summary: Update APME portal settings
      description: Persists the portal-wide default ansible-core scan target.
      tags:
        - APME
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetAnsibleCoreVersion
              properties:
                targetAnsibleCoreVersion:
                  type: string
                  description: ansible-core minor version (e.g. 2.17)
      responses:
        '200':
          description: Updated APME settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  enableAi:
                    type: boolean
                  publishViaGateway:
                    type: boolean
                  targetAnsibleCoreVersion:
                    type: string
        '400':
          description: Invalid or unsupported ansible-core version
...
  /apme/repos/branch-check:
    get:
      operationId: checkApmeRepoBranch
      summary: Validate repository branch exists
      description: Verifies that the given repository URL and branch are reachable using portal SCM credentials.
      tags:
        - APME
      parameters:
        - name: repo_url
          in: query
          required: true
          schema:
            type: string
        - name: branch
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Branch is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
        '400':
          description: Missing or invalid query parameters
  /apme/projects:
    get:
      operationId: getApmeProjects
      summary: List all APME projects
      description: Returns all Ansible content projects registered with APME for policy analysis and modernization.
      tags:
        - APME
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: createApmeProject
      summary: Create a new APME project
      description: Registers a new Ansible content repository with APME for policy analysis and modernization tracking.
      tags:
        - APME
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                type: object
  /apme/projects/{projectId}:
    get:
      operationId: getApmeProject
      summary: Get a specific APME project
      description: Returns details of a specific APME project including health score, violation counts, and scan history.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
    delete:
      operationId: deleteApmeProject
      summary: Delete an APME project
      description: Removes a project from APME tracking. This does not delete the underlying repository.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Project deleted
  /apme/projects/{projectId}/scan-target:
    get:
      operationId: getApmeProjectScanTarget
      summary: Get effective ansible-core scan target for a project
      description: Resolves the ansible-core version used for quality scans, including project override and portal defaults.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resolved scan target
          content:
            application/json:
              schema:
                type: object
                properties:
                  effective:
                    type: string
                  source:
                    type: string
                    enum: [project, global, config, default]
                  globalDefault:
                    type: string
                  projectOverride:
                    type: string
    put:
      operationId: updateApmeProjectScanTarget
      summary: Set per-project ansible-core scan target override
      description: Persists a project-specific ansible-core scan target, or clears the override when null.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetAnsibleCoreVersion:
                  type: string
                  nullable: true
                  description: ansible-core minor version (e.g. 2.17), or null to clear project override
      responses:
        '200':
          description: Updated scan target resolution
          content:
            application/json:
              schema:
                type: object
                properties:
                  effective:
                    type: string
                  source:
                    type: string
                    enum: [project, global, config, default]
                  globalDefault:
                    type: string
                  projectOverride:
                    type: string
        '400':
          description: Invalid or unsupported ansible-core version
  /apme/projects/{projectId}/violations:
    get:
      operationId: getApmeViolations
      summary: Get violations for a project
      description: Returns all policy violations detected in the project from the most recent scan, including severity levels and file locations.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of violations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /apme/projects/{projectId}/dependencies:
    get:
      operationId: getApmeProjectDependencies
      summary: Get project dependencies from latest scan
      description: Returns Ansible collections and Python packages detected during the latest gateway scan for the project.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Collections and Python packages discovered in the latest scan
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      type: object
                  python_packages:
                    type: array
                    items:
                      type: object
  /apme/projects/{projectId}/operation:
    post:
      operationId: triggerApmeOperation
      summary: Trigger a scan operation
      description: Initiates a new policy scan on the project. Returns an operation ID for tracking progress.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Operation started
          content:
            application/json:
              schema:
                type: object
  /apme/projects/{projectId}/operation/state:
    get:
      operationId: getApmeOperationState
      summary: Get current operation state
      description: Returns the current state of an in-progress operation including phase and progress details.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operation state
          content:
            application/json:
              schema:
                type: object
        '404':
          description: No active operation
  /apme/projects/{projectId}/operation/approve:
    post:
      operationId: approveApmeProposals
      summary: Approve remediation proposals
      description: Approves selected remediation proposals from the current operation to be applied to the codebase.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Proposals approved
          content:
            application/json:
              schema:
                type: object
  /apme/projects/{projectId}/remediate:
    post:
      operationId: triggerApmeRemediation
      summary: Trigger remediation
      description: Initiates automated remediation of detected violations using AI-assisted code fixes.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Remediation started
          content:
            application/json:
              schema:
                type: object
  /apme/projects/{projectId}/activity:
    get:
      operationId: getApmeActivity
      summary: Get project scan history
      description: Returns the history of scans and remediation operations performed on the project.
      tags:
        - APME
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activity history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /apme/activity/{activityId}:
    get:
      operationId: getApmeActivityDetail
      summary: Get scan run detail
      description: Returns violations, AI proposals, and logs for a single scan or remediation run.
      tags:
        - APME
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activity detail
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Activity not found
...
  /apme/rules/{ruleId}/config:
    put:
      operationId: updateApmeRuleConfig
      summary: Update rule override config
      description: Applies portal-side overrides for rule severity, enabled state, or enforcement.
      tags:
        - APME
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated rule
    delete:
      operationId: deleteApmeRuleConfig
      summary: Reset rule override config
      description: Clears portal overrides so the rule reverts to gateway catalog defaults.
      tags:
        - APME
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Override removed
...
  /apme/suppressions/{suppressionId}:
    delete:
      operationId: deleteApmeSuppression
      summary: Remove a suppression
      description: Deletes a suppression so the violation can appear in quality results again.
      tags:
        - APME
      parameters:
        - name: suppressionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...